// src/engines/particle-sphere/particleSphere.types.ts
export type ParticleSphereConfig = {
  orbitSpeed: number;
  energyResponse: number;
  particleDensity: number;
  glowIntensity: number;
  spaceScale: number;
  colorSaturation: number;
  sparkleDensity: number;
  warmth: number;
  primaryColor: string;
  accentColor: string;
  showTitle: boolean;
};